+++ lib/graphql_rails/attributes/input_attribute.rb
@@ -7,19 +7,11 @@ module GraphqlRails
       require_relative './input_type_parser'
       require_relative './attribute_name_parser'
       include Attributable
+      include AttributeConfigurable
 
-      attr_reader :description
-
-      # rubocop:disable Metrics/ParameterLists
-      def initialize(name, type: nil, description: nil, subtype: nil, required: nil, options: {})
+      def initialize(name)
         @initial_name = name
-        @initial_type = type
-        @description = description
-        @options = options
-        @subtype = subtype
-        @required = required
       end
-      # rubocop:enable Metrics/ParameterLists
 
       def input_argument_args
         type = raw_input_type || input_type_parser.input_type_arg
@@ -31,13 +23,27 @@ module GraphqlRails
         { required: required?, description: description, camelize: false }
       end
 
+      def subtype(new_value = NOT_SET)
+        return @subtype if new_value == NOT_SET
+
+        @subtype = new_value
+        self
+      end
+
+      def enum(*enum_values)
+        return @enum if enum_values.empty?
+
+        @enum = enum_values.flatten
+        self
+      end
+
       def paginated?
         false
       end
 
       private
 
-      attr_reader :initial_name, :initial_type, :options, :subtype
+      attr_reader :initial_name
 
       def attribute_name_parser
         @attribute_name_parser ||= AttributeNameParser.new(
@@ -51,14 +57,23 @@ module GraphqlRails
 
       def input_type_parser
         @input_type_parser ||= begin
-          initial_parseable_type = initial_type || attribute_name_parser.graphql_type
+          initial_parseable_type = type || enum_type || attribute_name_parser.graphql_type
           InputTypeParser.new(initial_parseable_type, subtype: subtype)
         end
       end
 
+      def enum_type
+        return if enum.blank?
+
+        BuildEnumType.call(
+          "#{name}_#{attribute_name}_enum",
+          allowed_values: enum
+        )
+      end
+
       def raw_input_type
-        return initial_type if initial_type.is_a?(GraphQL::InputObjectType)
-        return initial_type.graphql_input_type if initial_type.is_a?(Model::Input)
+        return type if type.is_a?(GraphQL::InputObjectType)
+        return type.graphql_input_type if type.is_a?(Model::Input)
       end
     end
   end